Offload more details into the SimpleShaderProgram I have a sickness with abstraction, truly
